feat: @readymade/dom package
