+++ packages/@readymade/core/fesm2015/core.js
@@ -200,9 +200,6 @@ class NodeTree {
             id,
             node: clone
         };
-        if (key && value) {
-            this.updateNode(node, key, value);
-        }
         node.$init = true;
         return this.$flatMap[id];
     }
@@ -417,12 +414,6 @@ function getElementIndex(el) {
     return getSiblings(el).indexOf(el);
 }
 
-function __decorate(decorators, target, key, desc) {
-    var c = arguments.length, r = c < 3 ? target : desc === null ? desc = Object.getOwnPropertyDescriptor(target, key) : desc, d;
-    if (typeof Reflect === "object" && typeof Reflect.decorate === "function") r = Reflect.decorate(decorators, target, key, desc);
-    else for (var i = decorators.length - 1; i >= 0; i--) if (d = decorators[i]) r = (c < 3 ? d(r) : c > 3 ? d(target, key, r) : d(target, key)) || r;
-    return c > 3 && r && Object.defineProperty(target, key, r), r;
-}
 function __awaiter(thisArg, _arguments, P, generator) {
     function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
     return new (P || (P = Promise))(function (resolve, reject) {
@@ -571,6 +562,20 @@ function Listen(eventName, channelName) {
     };
 }
 
+function bindListeners(target) {
+    for (const prop in target) {
+        if (prop.includes(LISTEN_KEY)) {
+            this[prop].onListener.call(this);
+        }
+    }
+}
+function bindEmitters(target) {
+    for (const prop in target) {
+        if (prop.includes(EMIT_KEY)) {
+            target[prop].call(target);
+        }
+    }
+}
 class StructuralElement extends HTMLElement {
     constructor() {
         super();
@@ -616,1088 +621,5 @@ class CustomElement extends HTMLElement {
         }
     }
 }
-class AllCollectionComponent extends HTMLAllCollection {
-    constructor() {
-        super();
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class AnchorComponent extends HTMLAnchorElement {
-    constructor() {
-        super();
-        attachDOM(this);
-        attachStyle(this);
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class AreaComponent extends HTMLAreaElement {
-    constructor() {
-        super();
-        attachStyle(this);
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class AudioComponent extends HTMLAudioElement {
-    constructor() {
-        super();
-        attachStyle(this);
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class BRComponent extends HTMLBRElement {
-    constructor() {
-        super();
-        attachStyle(this);
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class BodyComponent extends HTMLBodyElement {
-    constructor() {
-        super();
-        attachShadow(this, { mode: this.elementMeta.mode || 'open' });
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class ButtonComponent extends HTMLButtonElement {
-    constructor() {
-        super();
-        attachDOM(this);
-        attachStyle(this);
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class CanvasComponent extends HTMLCanvasElement {
-    constructor() {
-        super();
-        attachStyle(this);
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class CollectionComponent extends HTMLCollection {
-    constructor() {
-        super();
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class DListComponent extends HTMLDListElement {
-    constructor() {
-        super();
-        attachDOM(this);
-        attachStyle(this);
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class DataComponent extends HTMLDataElement {
-    constructor() {
-        super();
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class DetailsComponent extends HTMLDetailsElement {
-    constructor() {
-        super();
-        attachDOM(this);
-        attachStyle(this);
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class DivComponent extends HTMLDivElement {
-    constructor() {
-        super();
-        attachShadow(this, { mode: this.elementMeta.mode || 'open' });
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class EmbedComponent extends HTMLEmbedElement {
-    constructor() {
-        super();
-        attachStyle(this);
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class FieldSetComponent extends HTMLFieldSetElement {
-    constructor() {
-        super();
-        attachStyle(this);
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class FormControlsComponent extends HTMLFormControlsCollection {
-    constructor() {
-        super();
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class FormComponent extends HTMLFormElement {
-    constructor() {
-        super();
-        attachStyle(this);
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class HRComponent extends HTMLHRElement {
-    constructor() {
-        super();
-        attachStyle(this);
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class HeadComponent extends HTMLHeadElement {
-    constructor() {
-        super();
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class HeadingComponent extends HTMLHeadingElement {
-    constructor() {
-        super();
-        attachShadow(this, { mode: this.elementMeta.mode || 'open' });
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class HtmlComponent extends HTMLHtmlElement {
-    constructor() {
-        super();
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class IFrameComponent extends HTMLIFrameElement {
-    constructor() {
-        super();
-        attachStyle(this);
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class ImageComponent extends HTMLImageElement {
-    constructor() {
-        super();
-        attachStyle(this);
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class InputComponent extends HTMLInputElement {
-    constructor() {
-        super();
-        attachStyle(this);
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class LIComponent extends HTMLLIElement {
-    constructor() {
-        super();
-        attachDOM(this);
-        attachStyle(this);
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class LabelComponent extends HTMLLabelElement {
-    constructor() {
-        super();
-        attachDOM(this);
-        attachStyle(this);
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class LegendComponent extends HTMLLegendElement {
-    constructor() {
-        super();
-        attachStyle(this);
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class LinkComponent extends HTMLLinkElement {
-    constructor() {
-        super();
-        attachStyle(this);
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class MapComponent extends HTMLMapElement {
-    constructor() {
-        super();
-        attachDOM(this);
-        attachStyle(this);
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class MediaComponent extends HTMLMediaElement {
-    constructor() {
-        super();
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class MenuComponent extends HTMLMenuElement {
-    constructor() {
-        super();
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class MetaComponent extends HTMLMetaElement {
-    constructor() {
-        super();
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class MeterComponent extends HTMLMeterElement {
-    constructor() {
-        super();
-        attachStyle(this);
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class ModComponent extends HTMLModElement {
-    constructor() {
-        super();
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class OListComponent extends HTMLOListElement {
-    constructor() {
-        super();
-        attachDOM(this);
-        attachStyle(this);
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class ObjectComponent extends HTMLObjectElement {
-    constructor() {
-        super();
-        attachStyle(this);
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class OptGroupComponent extends HTMLOptGroupElement {
-    constructor() {
-        super();
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class OptionComponent extends HTMLOptionElement {
-    constructor() {
-        super();
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class OptionsCollectionComponent extends HTMLOptionsCollection {
-    constructor() {
-        super();
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class OutputComponent extends HTMLOutputElement {
-    constructor() {
-        super();
-        attachStyle(this);
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class ParagraphComponent extends HTMLParagraphElement {
-    constructor() {
-        super();
-        attachShadow(this, { mode: this.elementMeta.mode || 'open' });
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class ParamComponent extends HTMLParamElement {
-    constructor() {
-        super();
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class PictureComponent extends HTMLPictureElement {
-    constructor() {
-        super();
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class PreComponent extends HTMLPreElement {
-    constructor() {
-        super();
-        attachDOM(this);
-        attachStyle(this);
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class ProgressComponent extends HTMLProgressElement {
-    constructor() {
-        super();
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class QuoteComponent extends HTMLQuoteElement {
-    constructor() {
-        super();
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class ScriptComponent extends HTMLScriptElement {
-    constructor() {
-        super();
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class SelectComponent extends HTMLSelectElement {
-    constructor() {
-        super();
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class SlotComponent extends HTMLSlotElement {
-    constructor() {
-        super();
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class SourceComponent extends HTMLSourceElement {
-    constructor() {
-        super();
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class SpanComponent extends HTMLSpanElement {
-    constructor() {
-        super();
-        attachShadow(this, { mode: this.elementMeta.mode || 'open' });
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class StyleComponent extends HTMLStyleElement {
-    constructor() {
-        super();
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class TableCaptionComponent extends HTMLTableCaptionElement {
-    constructor() {
-        super();
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class TableCellComponent extends HTMLTableCellElement {
-    constructor() {
-        super();
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class TableColComponent extends HTMLTableColElement {
-    constructor() {
-        super();
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class TableComponent extends HTMLTableElement {
-    constructor() {
-        super();
-        attachDOM(this);
-        attachStyle(this);
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class TableRowComponent extends HTMLTableRowElement {
-    constructor() {
-        super();
-        attachDOM(this);
-        attachStyle(this);
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class TableSectionComponent extends HTMLTableSectionElement {
-    constructor() {
-        super();
-        attachDOM(this);
-        attachStyle(this);
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class TemplateComponent extends HTMLTemplateElement {
-    constructor() {
-        super();
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class TimeComponent extends HTMLTimeElement {
-    constructor() {
-        super();
-        attachStyle(this);
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class TitleComponent extends HTMLTitleElement {
-    constructor() {
-        super();
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class TrackComponent extends HTMLTrackElement {
-    constructor() {
-        super();
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class UListComponent extends HTMLUListElement {
-    constructor() {
-        super();
-        attachDOM(this);
-        attachStyle(this);
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class UnknownComponent extends HTMLUnknownElement {
-    constructor() {
-        super();
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
-class VideoComponent extends HTMLVideoElement {
-    constructor() {
-        super();
-        attachStyle(this);
-        if (this.bindEmitters) {
-            this.bindEmitters();
-        }
-        if (this.bindListeners) {
-            this.bindListeners();
-        }
-        if (this.onInit) {
-            this.onInit();
-        }
-    }
-}
 
-function changeNode(protoNode, key, regex, value, index) {
-    const node = document.importNode(protoNode, true);
-    let attr = '';
-    node.removeAttribute('repeat');
-    if (protoNode.textContent.startsWith(`{{${key}`)) {
-        const path = stripTemplateString(protoNode.textContent);
-        const template = path.substring(path.search(DOT_BRACKET_NOTATION_REGEX));
-        node.textContent = protoNode.textContent.replace(protoNode.textContent, isObject(value) ? findValueByString(value, template) : value);
-    }
-    for (const attribute of protoNode.attributes) {
-        attr = attribute.nodeName || attribute.name;
-        if (attr !== 'repeat') {
-            if (attr.includes('attr.')) {
-                if (!protoNode.getAttribute(attr.replace('attr.', ''))) {
-                    if (attribute.nodeName) {
-                        attr = attribute.nodeName.replace('attr.', '');
-                    }
-                    else if (attribute.name) {
-                        attr = attribute.name.replace('attr.', '');
-                    }
-                    if (!protoNode.setAttribute) {
-                        protoNode.setAttribute = function (i, v) { };
-                    }
-                    const path = stripTemplateString(attribute.nodeValue);
-                    const template = path.substring(path.search(DOT_BRACKET_NOTATION_REGEX));
-                    protoNode.setAttribute(attr, isObject(value) ? findValueByString(value, template) : value);
-                    const remove = attribute.nodeName || attribute.name;
-                    node.removeAttribute(remove);
-                }
-            }
-            const attributeValue = attribute.nodeValue || attribute.value;
-            if (attributeValue.startsWith(`{{${key}`)) {
-                if (!node.setAttribute) {
-                    node.setAttribute = function (i, v) { };
-                }
-                const path = stripTemplateString(attributeValue);
-                const template = path.substring(path.search(DOT_BRACKET_NOTATION_REGEX));
-                node.setAttribute(attr, attributeValue.replace(attributeValue, isObject(value) ? findValueByString(value, template) : value));
-            }
-            const check = attribute.nodeName || attribute.name;
-            if (check.includes('attr.')) {
-                node.removeAttribute(check);
-            }
-        }
-    }
-    protoNode.parentNode.appendChild(node);
-}
-function renderTemplate(elem, template, items) {
-    const bound = items.match(/(\w*)(?:\s)(?:of)(?:\s)(?:\{\{(?:\s*)(.*?)(?:\s*)\}\})/);
-    if (bound && bound.length) {
-        return;
-    }
-    if (!elem.parentNode) {
-        return;
-    }
-    const key = items.split('of')[0].trim();
-    const model = JSON.parse(items.split('of')[1].trim());
-    const regex = templateRegExp(key);
-    const clone = template.content.cloneNode(true);
-    const protoNode = clone.querySelector(`[repeat="${key}"]`);
-    if (Array.isArray(model)) {
-        for (let index = 0; index < model.length; index++) {
-            changeNode(protoNode, key, regex, model[index]);
-        }
-    }
-    protoNode.parentNode.removeChild(protoNode);
-    if (elem instanceof TemplateRepeater) {
-        elem.appendChild(clone);
-    }
-    else {
-        elem.parentNode.appendChild(clone);
-    }
-}
-let Repeater = class Repeater extends TemplateComponent {
-    constructor() {
-        super();
-        this.bindTemplate();
-    }
-    static get observedAttributes() {
-        return ['items'];
-    }
-    attributeChangedCallback(name, prev, next) {
-        switch (name) {
-            case 'items':
-                this.render(next);
-                break;
-        }
-    }
-    render(items) {
-        renderTemplate(this, this, items);
-    }
-};
-Repeater = __decorate([
-    Component({
-        selector: 'r-repeat',
-        custom: { extends: 'template' }
-    })
-], Repeater);
-let TemplateRepeater = class TemplateRepeater extends PseudoElement {
-    constructor() {
-        super();
-    }
-    static get observedAttributes() {
-        return ['items', 'template'];
-    }
-    attributeChangedCallback(name, prev, next) {
-        switch (name) {
-            case 'template':
-                this.setTemplate(next);
-                break;
-            case 'items':
-                this.setItems(next);
-                break;
-        }
-    }
-    setItems(items) {
-        this.$items = items;
-        this.render();
-    }
-    render() {
-        const template = document.querySelector(`[id="${this.$templateId}"]`);
-        if (template && this.$items) {
-            renderTemplate(this, template, this.$items);
-        }
-    }
-    setTemplate(id) {
-        this.$templateId = id;
-        this.render();
-    }
-};
-TemplateRepeater = __decorate([
-    Component({
-        selector: 'r-repeatr'
-    })
-], TemplateRepeater);
-
-export { AllCollectionComponent, AnchorComponent, AreaComponent, AudioComponent, BRComponent, BodyComponent, ButtonComponent, CanvasComponent, CollectionComponent, Component, CustomElement, DListComponent, DataComponent, DetailsComponent, DivComponent, EMIT_KEY, EmbedComponent, Emitter, EventDispatcher, FieldSetComponent, FormComponent, FormControlsComponent, HRComponent, HeadComponent, HeadingComponent, HtmlComponent, IFrameComponent, ImageComponent, InputComponent, LIComponent, LISTEN_KEY, LabelComponent, LegendComponent, LinkComponent, Listen, MapComponent, MediaComponent, MenuComponent, MetaComponent, MeterComponent, ModComponent, OListComponent, ObjectComponent, OptGroupComponent, OptionComponent, OptionsCollectionComponent, OutputComponent, ParagraphComponent, ParamComponent, PictureComponent, PreComponent, ProgressComponent, PseudoElement, QuoteComponent, Repeater, ScriptComponent, SelectComponent, SlotComponent, SourceComponent, SpanComponent, State, StructuralElement, StyleComponent, TableCaptionComponent, TableCellComponent, TableColComponent, TableComponent, TableRowComponent, TableSectionComponent, TemplateComponent, TemplateRepeater, TimeComponent, TitleComponent, TrackComponent, UListComponent, UnknownComponent, VideoComponent, attachDOM, attachShadow, attachStyle, bindTemplate, compileTemplate, css, define, getChildNodes, getElementIndex, getParent, getSiblings, html, noop, querySelector, querySelectorAll };
+export { Component, CustomElement, EMIT_KEY, Emitter, EventDispatcher, LISTEN_KEY, Listen, PseudoElement, State, StructuralElement, attachDOM, attachShadow, attachStyle, bindEmitters, bindListeners, bindTemplate, compileTemplate, css, define, getChildNodes, getElementIndex, getParent, getSiblings, html, noop, querySelector, querySelectorAll };
