+++ packages/@readymade/core/types/index.d.ts
@@ -1,5 +1,4 @@
-export * from './event/event';
-export * from './element/element';
-export * from './decorator/decorator';
-export * from './component/component';
-export * from './component/template';
+export * from './event';
+export * from './element';
+export * from './decorator';
+export * from './component';
