+++ src/client/index.ts
@@ -2,5 +2,5 @@ import { RdRouter, routing } from './app/routing';
 
 window['clientRouter'] = new RdRouter('#root', routing);
 
-export { Repeater, TemplateRepeater } from './../modules/core/component/template';
+export { Repeater, TemplateRepeater } from './../modules/dom/repeatr';
 export * from './app';
