+++ src/modules/dom/index.ts
@@ -1,0 +1,2 @@
+export * from './custom';
+export * from './repeatr';
