+++ src/modules/dom/package.json
@@ -1,0 +1,30 @@
+{
+  "name": "@readymade/dom",
+  "version": "2.0.0-beta.0",
+  "description": "JavaScript microlibrary for developing Web Components with TypeScript and Decorators",
+  "umd": "./bundles/dom.js",
+  "main": "./fesm2015/dom.js",
+  "es2015": "./fesm2015/dom.js",
+  "fesm2015": "./fesm2015/dom.js",
+  "types": "./types/index.d.ts",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/readymade-ui/readymade.git"
+  },
+  "keywords": [
+    "customelements",
+    "webcomponents",
+    "typescript",
+    "decorators",
+    "javascript"
+  ],
+  "author": "Steve Belovarich",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/readymade-ui/readymade/issues"
+  },
+  "homepage": "https://github.com/readymade-ui/readymade#readme"
+}
