+++ src/modules/dom/tsconfig.json
@@ -1,0 +1,19 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "baseUrl": "./",
+    "experimentalDecorators": true,
+    "moduleResolution": "node",
+    "module": "ES2015",
+    "target": "ES2015",
+    "sourceMap": false,
+    "typeRoots": ["node_modules/@types"],
+    "lib": ["es2017", "dom", "dom.iterable"],
+    "declaration": true,
+    "declarationDir": "typings",
+    "emitDeclarationOnly": true,
+  },
+  "include": [
+    "./**/*.ts"
+  ]
+}
